requirements.txt
langchain==0.3.8
langchain-openai==0.2.9
langchain-anthropic==0.3.0
python-dotenv==1.0.1
twitter-api-client==0.10.22
tweepy==4.14.0
langchain_community==0.3.8
alembic==1.14.0
psycopg2==2.9.10
sqlalchemy==2.0.35
python-dateutil==2.9.0.post0
python-telegram-bot==21.8
pytz==2024.2
aiogram==3.15.0
